refactor: move methods to hooks
