+++ components/Screens/ProfileScreen.js
@@ -5,6 +5,9 @@ import { TextInput, TouchableOpacity, View, Alert } from 'react-native';
 
 import { styles } from '../../assets/styles/ProfileScreenStyles';
 import { NanumText } from '../StyledText';
+import useResetPassword from '../../hooks/useResetPassword';
+import useLogOut from '../../hooks/useLogOut';
+import useUpdateProfile from '../../hooks/useUpdateProfile';
 
 export default function Profile() {
   const database = firebase.database();
@@ -18,118 +21,11 @@ export default function Profile() {
   const [failureEmailMessage, setFailureEmailMessage] = React.useState(false);
   const [timeoutMessage, setTimeOutMessage] = React.useState(false);
 
-  const resetPassword = () => {
-    firebase.auth().sendPasswordResetEmail(user.email)
-      .then(() => {
-        setSuccessEmailMessage(true);
-        setTimeout(() => {
-          setSuccessEmailMessage(false);
-        }, 2000);
-      })
-      .catch(() => {
-        setFailureEmailMessage(true);
-        setTimeout(() => {
-          setFailureEmailMessage(false);
-        }, 2000);
-      })
-  };
+  const resetPassword = useResetPassword(email, setSuccessEmailMessage, setFailureEmailMessage);
 
-  const updateProfile = () => {
-    const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  const updateProfile = useUpdateProfile(emailAddress, user, name, database, setSuccessMessage, setFailureMessage, setTimeOutMessage);
 
-    if (emailAddress !== user.email) {
-      if (emailRegex.test(emailAddress)) {
-        user.updateEmail(emailAddress)
-          .then(() => {
-            user.updateProfile({ displayName: name })
-              .then(() => {
-                database.ref('users/' + user.uid).update({
-                  email: emailAddress,
-                  name: name,
-                  newUser: false,
-                  plan: 'Free',
-                  uid: id
-                });
-
-                database.ref('users/' + user.uid + '/closet/0').update({
-                  name: name
-                });
-
-                setSuccessMessage(true);
-                setTimeout(() => {
-                  setSuccessMessage(false);
-                }, 2000);
-              })
-              .catch(() => {
-                setFailureMessage(true);
-                setTimeout(() => {
-                  setFailureMessage(false);
-                }, 2000);
-              })
-          })
-          .catch(err => {
-            if (err.code.includes('email-already-in-use')) {
-              Alert.alert(
-                'Invalid Email Address',
-                'Email address already in use',
-                [
-                  { text: 'Return to Profile' }
-                ]
-              )
-            } else {
-              setTimeOutMessage(true);
-              setTimeout(() => {
-                setTimeOutMessage(false);
-              }, 4000);
-            };
-          });
-      } else {
-        Alert.alert(
-          'Invalid Email',
-          'Please enter a valid email address',
-          [
-            { text: 'Return to Profile' }
-          ]
-        )
-      };
-    } else {
-      user.updateProfile({
-        displayName: name
-      }).then(() => {
-        database.ref('users/' + user.uid).update({
-          email: emailAddress,
-          name: name,
-          newUser: false,
-          plan: 'Free',
-          uid: id
-        });
-
-        database.ref('users/' + user.uid + '/closet/0').update({
-          name: name
-        });
-
-        setSuccessMessage(true);
-        setTimeout(() => {
-          setSuccessMessage(false);
-        }, 2000);
-      }).catch(() => {
-        setFailureMessage(true);
-        setTimeout(() => {
-          setFailureMessage(false);
-        }, 2000);
-      });
-    };
-  };
-
-  const logout = () => {
-    firebase.auth().signOut()
-      .then(() => {
-        console.log('Successfully signed out!')
-      })
-      .catch(err => {
-        console.log(err.code, err.message);
-      });
-  };
+  const logout = useLogOut();
 
   return (
     <View style={styles.container}>
@@ -176,23 +72,19 @@ export default function Profile() {
         <View style={styles.successMessage}>
           <NanumText style={{ color:'#6674DE' }}>Successfully updated profile!</NanumText>
         </View>
-      ) : null}
-      {failureMessage ? (
+      ) : failureMessage ? (
         <View style={styles.failureMessage}>
           <NanumText style={{ color:'white' }}>Unable to update profile</NanumText>
         </View>
-      ) : null}
-      {successEmailMessage ? (
+      ) : successEmailMessage ? (
         <View style={styles.successMessage}>
           <NanumText style={{ color:'#6674DE' }}>Password reset email sent to {user.email}!</NanumText>
         </View>
-      ) : null}
-      {failureEmailMessage ? (
+      ) : failureEmailMessage ? (
         <View style={styles.failureMessage}>
           <NanumText style={{ color:'white' }}>There was an error sending the email. Please try again</NanumText>
         </View>
-      ) : null}
-      {timeoutMessage ? (
+      ) : timeoutMessage ? (
         <View style={styles.failureMessage}>
           <NanumText style={{ color:'white' }}>To change your email address, please logout and log back in to verify this is your account</NanumText>
         </View>
