+++ hooks/useUpdateProfile.js
@@ -1,0 +1,90 @@
+import * as React from 'react';
+
+import { Alert } from 'react-native';
+
+export default function useUpdateProfile({ emailAddress, user, name, database, setSuccessMessage, setFailureMessage, setTimeOutMessage }) {
+  const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
+  if (emailAddress !== user.email) {
+    if (emailRegex.test(emailAddress)) {
+      user.updateEmail(emailAddress)
+        .then(() => {
+          user.updateProfile({ displayName: name })
+            .then(() => {
+              database.ref('users/' + user.uid).update({
+                email: emailAddress,
+                name: name,
+                newUser: false,
+                plan: 'Free',
+                uid: id
+              });
+
+              database.ref('users/' + user.uid + '/closet/0').update({
+                name: name
+              });
+
+              setSuccessMessage(true);
+              setTimeout(() => {
+                setSuccessMessage(false);
+              }, 2000);
+            })
+            .catch(() => {
+              setFailureMessage(true);
+              setTimeout(() => {
+                setFailureMessage(false);
+              }, 2000);
+            })
+        })
+        .catch(err => {
+          if (err.code.includes('email-already-in-use')) {
+            Alert.alert(
+              'Invalid Email Address',
+              'Email address already in use',
+              [
+                { text: 'Return to Profile' }
+              ]
+            )
+          } else {
+            setTimeOutMessage(true);
+            setTimeout(() => {
+              setTimeOutMessage(false);
+            }, 4000);
+          };
+        });
+    } else {
+      Alert.alert(
+        'Invalid Email',
+        'Please enter a valid email address',
+        [
+          { text: 'Return to Profile' }
+        ]
+      )
+    };
+  } else {
+    user.updateProfile({
+      displayName: name
+    }).then(() => {
+      database.ref('users/' + user.uid).update({
+        email: emailAddress,
+        name: name,
+        newUser: false,
+        plan: 'Free',
+        uid: id
+      });
+
+      database.ref('users/' + user.uid + '/closet/0').update({
+        name: name
+      });
+
+      setSuccessMessage(true);
+      setTimeout(() => {
+        setSuccessMessage(false);
+      }, 2000);
+    }).catch(() => {
+      setFailureMessage(true);
+      setTimeout(() => {
+        setFailureMessage(false);
+      }, 2000);
+    });
+  };
+}
